Added silk into development configuration by default.
