+++ cvat/requirements/development.txt
@@ -6,3 +6,4 @@ pylint-plugin-utils==0.7
 rope==0.17.0
 django-extensions==3.0.8
 snakeviz==2.1.0
+django-silk==5.0.3
