--- cvat/requirements/staging.txt
@@ -1,2 +1,0 @@
--r production.txt
-django-silk==5.0.3
